+++ lib/solargraph/source_map/mapper.rb
@@ -12,7 +12,7 @@ module Solargraph
 
       private_class_method :new
 
-      MACRO_REGEXP = /(@\!method|@\!attribute|@\!domain|@\!macro|@\!parse|@\!override)/.freeze
+      MACRO_REGEXP = /(@\!method|@\!attribute|@\!visibility|@\!domain|@\!macro|@\!parse|@\!override)/.freeze
 
       # Generate the data.
       #
@@ -58,8 +58,10 @@ module Solargraph
         @pins ||= []
       end
 
+      # @param position [Solargraph::Position]
+      # @return [Solargraph::Pin::Closure]
       def closure_at(position)
-        @pins.select{|pin| pin.is_a?(Pin::Closure) and pin.location.range.contain?(position)}.last
+        pins.select{|pin| pin.is_a?(Pin::Closure) and pin.location.range.contain?(position)}.last
       end
 
       def process_comment source_position, comment_position, comment
@@ -149,6 +151,27 @@ module Solargraph
               pins.last.docstring.add_tag YARD::Tags::Tag.new(:param, '', pins.last.return_type.to_s.split(', '), 'value')
             end
           end
+        when 'visibility'
+          begin
+            kind = directive.tag.text&.to_sym
+            return unless [:private, :protected, :public].include?(kind)
+
+            name = directive.tag.name
+            closure = closure_at(source_position) || @pins.first
+            if closure.location.range.start.line < comment_position.line
+              closure = closure_at(comment_position)
+            end
+            if closure.is_a?(Pin::Method) && no_empty_lines?(comment_position.line, source_position.line)
+              # @todo Smelly instance variable access
+              closure.instance_variable_set(:@visibility, kind)
+            else
+              matches = pins.select{ |pin| pin.is_a?(Pin::Method) && pin.name == name && pin.namespace == namespace && pin.context.scope == namespace.is_a?(Pin::Singleton) ? :class : :instance }
+              matches.each do |pin|
+                # @todo Smelly instance variable access
+                pin.instance_variable_set(:@visibility, kind)
+              end
+            end
+          end
         when 'parse'
           begin
             ns = closure_at(source_position)
@@ -176,6 +199,10 @@ module Solargraph
         when 'override'
           pins.push Pin::Reference::Override.new(location, directive.tag.name, docstring.tags)
         end
+      end
+
+      def no_empty_lines?(line1, line2)
+        @code.lines[line1..line2].none? { |line| line.strip.empty? }
       end
 
       def remove_inline_comment_hashes comment