+++ spec/source_map/mapper_spec.rb
@@ -189,6 +189,75 @@ describe Solargraph::SourceMap::Mapper do
     expect(pin.return_type.tag).to eq('String')
   end
 
+  it "processes private visibility directives attached to methods" do
+    map = Solargraph::SourceMap.load_string(%(
+      class Foo
+        # @!visibility private
+        def bar
+        end
+      end
+    ))
+    expect(map.first_pin('Foo#bar').visibility).to be(:private)
+  end
+
+  it "processes protected visibility directives attached to methods" do
+    map = Solargraph::SourceMap.load_string(%(
+      class Foo
+        # @!visibility protected
+        def bar
+        end
+      end
+    ))
+    expect(map.first_pin('Foo#bar').visibility).to be(:protected)
+  end
+
+  it "processes public visibility directives attached to methods" do
+    map = Solargraph::SourceMap.load_string(%(
+      class Foo
+        # @!visibility public
+        def bar
+        end
+      end
+    ))
+    expect(map.first_pin('Foo#bar').visibility).to be(:public)
+  end
+
+  it "does not process attached visibility directives on other methods" do
+    map = Solargraph::SourceMap.load_string(%(
+      class Example
+        # @!visibility private
+        def method1; end
+
+        def method2; end
+      end
+    ))
+    method1 = map.first_pin('Example#method1')
+    expect(method1.visibility).to be(:private)
+    method2 = map.first_pin('Example#method2')
+    expect(method2.visibility).to be(:public)
+  end
+
+  it "processes class-wide private visibility directives" do
+    map = Solargraph::SourceMap.load_string(%(
+      class Example
+        # @!visibility private
+
+        def method1; end
+
+        def method2; end
+
+        # @!visibility public
+        def method3; end
+      end
+    ))
+    method1 = map.first_pin('Example#method1')
+    expect(method1.visibility).to be(:private)
+    method2 = map.first_pin('Example#method2')
+    expect(method2.visibility).to be(:private)
+    method3 = map.first_pin('Example#method3')
+    expect(method3.visibility).to be(:public)
+  end
+
   it "processes attribute directives at class endings" do
     map = Solargraph::SourceMap.load_string(%(
       class Foo